+++ Helper/Cart.php
@@ -61,7 +61,7 @@ use Bolt\Boltpay\Exception\BoltException;
 use Bolt\Boltpay\Model\ErrorResponse as BoltErrorResponse;
 use Bolt\Boltpay\Helper\MetricsClient;
 use Bolt\Boltpay\Helper\FeatureSwitch\Decider as DeciderHelper;
-use Magento\Catalog\Model\Config\Source\Product\Thumbnail as ThumbnailSource;;
+use Magento\Catalog\Model\Config\Source\Product\Thumbnail as ThumbnailSource;
 
 /**
  * Boltpay Cart helper
@@ -684,12 +684,24 @@ class Cart extends AbstractHelper
     protected function getImmutableQuoteIdFromBoltOrder($boltOrder)
     {
         $response = $boltOrder ? $boltOrder->getResponse() : null;
-        list(, $immutableQuoteId) = $response ? explode(' / ', $response->cart->display_id) : [null, null];
+        $immutableQuoteId = null;
+        // If response is null don't even bother trying, we return null
+        if ($response)
+        {
+            try
+            {
+                $immutableQuoteId = $response->cart->metadata->immutable_quote_id;
+            }
+            catch (\Exception $e)
+            {
+                $this->bugsnag->notifyException($e);
+            }
+        }
         return $immutableQuoteId;
     }
 
     /**
-     * Check if the quote is noot deleted
+     * Check if the quote is not deleted
      *
      * @param int|string $quoteId
      * @return bool
@@ -833,7 +845,7 @@ class Cart extends AbstractHelper
      */
     public function doesOrderExist($cart, $quote)
     {
-        list($incrementId,) = isset($cart['display_id']) ? explode(' / ', $cart['display_id']) : [null, null];
+        $incrementId = $cart['display_id'];
         $order = $this->getOrderByIncrementId($incrementId);
 
         if ($quote && !$order) {
@@ -1592,8 +1604,11 @@ class Cart extends AbstractHelper
         $billingAddress  = $immutableQuote->getBillingAddress();
         $shippingAddress = $immutableQuote->getShippingAddress();
 
-        //Use display_id to hold and transmit, all the way back and forth, both reserved order id and immutable quote id
-        $cart['display_id'] = $immutableQuote->getReservedOrderId() . ' / ' . $immutableQuote->getId();
+        //Use display_id to hold and transmit, all the way back and forth, reserved order id
+        $cart['display_id'] = $immutableQuote->getReservedOrderId();
+
+        //Store immutable quote id in metadata of cart
+        $cart['metadata']['immutable_quote_id'] = $immutableQuote->getId();
 
         //Currency
         $currencyCode = $immutableQuote->getQuoteCurrencyCode();
@@ -2208,10 +2223,10 @@ class Cart extends AbstractHelper
         }
 
         // get configured Product model getters that can restrict Bolt checkout usage
-        $productRestrictionMethods = $toggleCheckout->productRestrictionMethods ?: [];
+        $productRestrictionMethods = isset($toggleCheckout->productRestrictionMethods) ? $toggleCheckout->productRestrictionMethods : [];
 
         // get configured Quote Item getters that can restrict Bolt checkout usage
-        $itemRestrictionMethods = $toggleCheckout->itemRestrictionMethods ?: [];
+        $itemRestrictionMethods = isset($toggleCheckout->itemRestrictionMethods) ? $toggleCheckout->itemRestrictionMethods : [];
 
         if (!$productRestrictionMethods && !$itemRestrictionMethods) {
             return false;
@@ -2264,36 +2279,37 @@ class Cart extends AbstractHelper
         }
 
         //add item to quote
-        $item = $request['items'][0];
-        $product = $this->productRepository->getbyId($item['reference']);
+        foreach ($request['items'] as $item) {
+            $product = $this->productRepository->getbyId($item['reference']);
 
-        $options = json_decode($item['options'], true);
-        if (isset($options['storeId']) && $options['storeId']) {
-            $quote->setStoreId($options['storeId']);
-        }
-        unset($options['storeId']);
-        unset($options['form_key']);
-        $options['qty'] = $item['quantity'];
-        $options = new \Magento\Framework\DataObject($options);
-
-        try {
-            $quote->addProduct($product, $options);
-        } catch (\Exception $e) {
-            $error_message = $e->getMessage();
-            if ($error_message == 'Product that you are trying to add is not available.') {
-                throw new BoltException(
-                    __($error_message),
-                    null,
-                    BoltErrorResponse::ERR_PPC_OUT_OF_STOCK
-                );
-            } else {
-                throw new BoltException(
-                    __('The requested qty is not available'),
-                    null,
-                    BoltErrorResponse::ERR_PPC_INVALID_QUANTITY
-                );
+            $options = json_decode($item['options'], true);
+            if (isset($options['storeId']) && $options['storeId']) {
+                $quote->setStoreId($options['storeId']);
             }
-        };
+            unset($options['storeId']);
+            unset($options['form_key']);
+            $options['qty'] = $item['quantity'];
+            $options = new \Magento\Framework\DataObject($options);
+
+            try {
+                $quote->addProduct($product, $options);
+            } catch (\Exception $e) {
+                $error_message = $e->getMessage();
+                if ($error_message == 'Product that you are trying to add is not available.') {
+                    throw new BoltException(
+                        __($error_message),
+                        null,
+                        BoltErrorResponse::ERR_PPC_OUT_OF_STOCK
+                    );
+                } else {
+                    throw new BoltException(
+                        __('The requested qty is not available'),
+                        null,
+                        BoltErrorResponse::ERR_PPC_INVALID_QUANTITY
+                    );
+                }
+            };
+        }
 
         $quote->reserveOrderId();
 
@@ -2374,7 +2390,7 @@ class Cart extends AbstractHelper
             }
 
             // get immutable quote id stored with cart data
-            list(, $cartReference) = $response ? explode(' / ', $responseData['cart']['display_id']) : [null, ''];
+            $cartReference = $response ? $responseData['cart']['metadata']['immutable_quote_id'] : '';
 
             $cart = array_merge($responseData['cart'], [
                 'orderToken'    => $response ? $responseData['token'] : '',