+++ Helper/Discount.php
@@ -138,6 +138,11 @@ class Discount extends AbstractHelper
      * @var ThirdPartyModuleFactory
      */
     private $mirasvitStoreCreditCalculationConfig;
+    
+    /**
+     * @var ThirdPartyModuleFactory
+     */
+    protected $mirasvitStoreCreditCalculationConfigLegacy;
 
     /**
      * @var ThirdPartyModuleFactory
@@ -239,6 +244,7 @@ class Discount extends AbstractHelper
      * @param ThirdPartyModuleFactory $mirasvitStoreCreditHelper
      * @param ThirdPartyModuleFactory $mirasvitStoreCreditCalculationHelper
      * @param ThirdPartyModuleFactory $mirasvitStoreCreditCalculationConfig
+     * @param ThirdPartyModuleFactory $mirasvitStoreCreditCalculationConfigLegacy
      * @param ThirdPartyModuleFactory $mirasvitRewardsPurchaseHelper
      * @param ThirdPartyModuleFactory $mirasvitStoreCreditConfig
      * @param ThirdPartyModuleFactory $mageplazaGiftCardCollection
@@ -274,6 +280,7 @@ class Discount extends AbstractHelper
         ThirdPartyModuleFactory $mirasvitStoreCreditHelper,
         ThirdPartyModuleFactory $mirasvitStoreCreditCalculationHelper,
         ThirdPartyModuleFactory $mirasvitStoreCreditCalculationConfig,
+        ThirdPartyModuleFactory $mirasvitStoreCreditCalculationConfigLegacy,
         ThirdPartyModuleFactory $mirasvitStoreCreditConfig,
         ThirdPartyModuleFactory $mirasvitRewardsPurchaseHelper,
         ThirdPartyModuleFactory $mageplazaGiftCardCollection,
@@ -308,6 +315,7 @@ class Discount extends AbstractHelper
         $this->mirasvitStoreCreditHelper = $mirasvitStoreCreditHelper;
         $this->mirasvitStoreCreditCalculationHelper = $mirasvitStoreCreditCalculationHelper;
         $this->mirasvitStoreCreditCalculationConfig = $mirasvitStoreCreditCalculationConfig;
+        $this->mirasvitStoreCreditCalculationConfigLegacy = $mirasvitStoreCreditCalculationConfigLegacy;
         $this->mirasvitStoreCreditConfig = $mirasvitStoreCreditConfig;
         $this->mirasvitRewardsPurchaseHelper = $mirasvitRewardsPurchaseHelper;
         $this->mageplazaGiftCardCollection = $mageplazaGiftCardCollection;
@@ -802,6 +810,12 @@ class Discount extends AbstractHelper
         if (!$paymentOnly) {
             /** @var \Mirasvit\Credit\Api\Config\CalculationConfigInterface $miravitCalculationConfig */
             $miravitCalculationConfig = $this->mirasvitStoreCreditCalculationConfig->getInstance();
+            // For old version of Mirasvit Store Credit plugin,
+            // \Magento\Framework\ObjectManagerInterface can not create instance of \Mirasvit\Credit\Api\Config\CalculationConfigInterface properly,
+            // so we use \Mirasvit\Credit\Service\Config\CalculationConfig instead.
+            if (empty($miravitCalculationConfig)) {
+                $miravitCalculationConfig = $this->mirasvitStoreCreditCalculationConfigLegacy->getInstance();
+            }
             if ($miravitCalculationConfig->isTaxIncluded() || $miravitCalculationConfig->IsShippingIncluded()) {
                 return $miravitBalanceAmount;
             }
@@ -1293,8 +1307,8 @@ class Discount extends AbstractHelper
      */
     public function convertToBoltDiscountType($couponCode)
     {
-        if(empty($couponCode)) {
-            return '';
+        if ($couponCode == "") {
+            return "fixed_amount";
         }
         
         try {
@@ -1326,7 +1340,7 @@ class Discount extends AbstractHelper
                 return "shipping";
         }
 
-        return "";
+        return "fixed_amount";
     }
     
     /**