+++ Test/Unit/Model/Api/ShippingMethodsTest.php
@@ -20,7 +20,6 @@ namespace Bolt\Boltpay\Test\Unit\Model\Api;
 use Bolt\Boltpay\Exception\BoltException;
 use Bolt\Boltpay\Model\Api\ShippingMethods as BoltShippingMethods;
 use Bolt\Boltpay\Test\Unit\TestHelper;
-use Magento\Framework\Phrase;
 use Magento\Framework\Webapi\Exception as WebapiException;
 use PHPUnit\Framework\TestCase;
 use Bolt\Boltpay\Helper\Hook as HookHelper;
@@ -48,7 +47,7 @@ use Bolt\Boltpay\Helper\Discount as DiscountHelper;
 use Magento\SalesRule\Model\RuleFactory as RuleFactory;
 use Magento\SalesRule\Model\Rule;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
-
+use Magento\Framework\Serialize\Serializer\Serialize;
 /**
  * Class ShippingMethodsTest
  *
@@ -60,7 +59,7 @@ class ShippingMethodsTest extends TestCase
     const PARENT_QUOTE_ID = 1000;
     const IMMUTABLE_QUOTE_ID = 1001;
     const INCREMENT_ID = 100050001;
-    const DISPLAY_ID = self::INCREMENT_ID . ' / ' . self::IMMUTABLE_QUOTE_ID;
+    const DISPLAY_ID = self::INCREMENT_ID;
     const STORE_ID = 1;
 
     /**
@@ -184,6 +183,11 @@ class ShippingMethodsTest extends TestCase
     private $storeMock;
 
     /**
+     * @var Serialize
+     */
+    private $serialize;
+
+    /**
      * @inheritdoc
      */
     public function setUp()
@@ -276,6 +280,7 @@ class ShippingMethodsTest extends TestCase
         $this->priceHelper->method('currency')->willReturnArgument(0);
         $this->sessionHelper = $this->createMock(SessionHelper::class);
         $this->discountHelper = $this->createMock(DiscountHelper::class);
+        $this->serialize = $this->getMockBuilder(Serialize::class)->enableProxyingToOriginalMethods()->getMock();
         $this->ruleFactory = $this->getMockBuilder(RuleFactory::class)
             ->setMethods(
                 [
@@ -322,7 +327,10 @@ class ShippingMethodsTest extends TestCase
     public function getShippingMethods_emptyQuote()
     {
         $cart = [
-            'display_id' => self::DISPLAY_ID
+            'display_id' => self::DISPLAY_ID,
+            'metadata' => [
+                'immutable_quote_id' => self::IMMUTABLE_QUOTE_ID
+            ]
         ];
         $shippingAddress = [
             'street_address1' => 'test'
@@ -434,12 +442,18 @@ class ShippingMethodsTest extends TestCase
             // common case
             [[
                 'display_id'      => self::DISPLAY_ID,
-                'order_reference' => self::PARENT_QUOTE_ID
+                'order_reference' => self::PARENT_QUOTE_ID,
+                'metadata'        => [
+                    'immutable_quote_id' => self::IMMUTABLE_QUOTE_ID,
+                ],
             ],self::PARENT_QUOTE_ID],
             // product page checkout case
             [[
                 'display_id'      => self::DISPLAY_ID,
-                'order_reference' => self::IMMUTABLE_QUOTE_ID
+                'order_reference' => self::IMMUTABLE_QUOTE_ID,
+                'metadata'        => [
+                    'immutable_quote_id' => self::IMMUTABLE_QUOTE_ID,
+                ],
             ],self::IMMUTABLE_QUOTE_ID]
         ];
     }
@@ -540,7 +554,15 @@ class ShippingMethodsTest extends TestCase
         $e = new WebapiException(__('Precondition Failed'), 6001, 412);
         $this->currentMock->method('preprocessHook')->willThrowException($e);
         $this->expectErrorResponse($e->getCode(), $e->getMessage(), $e->getHttpCode());
-        $this->assertNull($this->currentMock->getShippingMethods(['display_id' => self::DISPLAY_ID], []));
+        $this->assertNull($this->currentMock->getShippingMethods(
+            [
+                'display_id' => self::DISPLAY_ID,
+                'metadata'   => [
+                    'immutable_quote_id' => self::IMMUTABLE_QUOTE_ID,
+                    ]
+            ],
+            [])
+        );
     }
 
     /**
@@ -556,7 +578,10 @@ class ShippingMethodsTest extends TestCase
                     'quantity' => '2',
                     'total_amount' => '60000'
                 ]
-            ]
+            ],
+            'metadata' => [
+                'immutable_quote_id' => self::IMMUTABLE_QUOTE_ID,
+            ],
         ];
         $shippingAddress = [
             'company'         => "",
@@ -1560,7 +1585,8 @@ Room 4000',
                 $this->priceHelper,
                 $this->sessionHelper,
                 $this->discountHelper,
-                $this->ruleFactory
+                $this->ruleFactory,
+                $this->serialize
                 ]
             )
             ->setMethods($methods);